TypeScript Support
Not sure how to handle `ghosted` and `return null` just yet.

The TSConfig is only to demo these types in the source file, to show that it works at the type-checking level. To use the types in TS with this plugin, you just have to add `babel-plugin-glowup-vibes` to your own TSConfig in the `types` config option, and use Babel to transpile your TypeScript down to JS. Or another option is that you could use `// @ts-check` if you don't want to use the full TS syntax.